+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - missing parameters to Boleto, BrcodePayment, Deposit, DictKey, Event, Invoice, Transfer and Workspace resources
 - Workspace.update() to allow parameter updates
 - Base exception class
+- Invoice::Payment sub-resource to allow retrieval of invoice payment information
 
 ### Fixed
 - special characters in brcodePreview query